+++ frictionless/plugins/csv/__init__.py
@@ -1,0 +1,3 @@
+from .plugin import CsvPlugin
+from .dialect import CsvDialect
+from .parser import CsvParser
